+++ services/users/app/get_user_by_id_test.go
@@ -52,11 +52,11 @@ func TestGetUserById_Do(t *testing.T) {
 
 		deps.mockCrudRepository.
 			EXPECT().
-			GetById(gomock.Any(), database.UsersCollection, id, gomock.Any()).
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
 			Times(1).
 			Return(mockExpectedError)
 
-		_, err := deps.getUserByIdImpl.Do(deps.ctx, id)
+		_, err := deps.getUserByIdImpl.Do(deps.ctx, &app.GetUserByIdInput{Id: id, Deleted: false})
 		if err == nil {
 			t.Fail()
 		}
@@ -70,11 +70,11 @@ func TestGetUserById_Do(t *testing.T) {
 
 		deps.mockCrudRepository.
 			EXPECT().
-			GetById(gomock.Any(), database.UsersCollection, id, gomock.Any()).
+			GetById(gomock.Any(), database.UsersCollection, id, false, gomock.Any()).
 			Times(1).
 			Return(nil)
 
-		_, err := deps.getUserByIdImpl.Do(deps.ctx, id)
+		_, err := deps.getUserByIdImpl.Do(deps.ctx, &app.GetUserByIdInput{Id: id, Deleted: false})
 		if err != nil {
 			log.Fatal(err)
 		}