+++ services/users/app/update_user_by_id.go
@@ -55,19 +55,21 @@ func (gu *UpdateUserByIdImpl) Do(
 ) (*UpdateUserByIdOutput, error) {
 	var existentUser domain.UserDatabaseNoPassword
 
-	err := gu.userRepository.GetByEmail(
-		ctx,
-		database.UsersCollection,
-		input.Email,
-		&existentUser,
-	)
+	if input.Email != "" {
+		err := gu.userRepository.GetByEmail(
+			ctx,
+			database.UsersCollection,
+			input.Email,
+			&existentUser,
+		)
 
-	if err != nil {
-		return nil, err
-	}
+		if err != nil {
+			return nil, err
+		}
 
-	if existentUser != (domain.UserDatabaseNoPassword{}) && existentUser.Id != id {
-		return nil, errors.New(exception.CodePermission)
+		if existentUser != (domain.UserDatabaseNoPassword{}) && existentUser.Id != id {
+			return nil, errors.New(exception.CodePermission)
+		}
 	}
 
 	if input.Password != "" {
@@ -84,8 +86,9 @@ func (gu *UpdateUserByIdImpl) Do(
 
 	input.UpdatedAt = time.Now()
 
-	output := UpdateUserByIdOutput{}
-	err = gu.crudRepository.UpdateById(ctx, database.UsersCollection, id, &input, &output)
+	var output = UpdateUserByIdOutput{}
+
+	err := gu.crudRepository.UpdateById(ctx, database.UsersCollection, id, &input, &output)
 
 	if err != nil {
 		return nil, err