+++ services/users/mocks/update_user_by_id_interface_mock.go
@@ -43,16 +43,16 @@ func (m *MockUpdateUserByIdInterface) EXPECT() *MockUpdateUserByIdInterfaceMockR
 }
 
 // Do mocks base method.
-func (m *MockUpdateUserByIdInterface) Do(ctx context.Context, id string, updateUser *app.UpdateUserByIdInput) (*app.UpdateUserByIdOutput, error) {
+func (m *MockUpdateUserByIdInterface) Do(ctx context.Context, id string, input *app.UpdateUserByIdInput) (*app.UpdateUserByIdOutput, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Do", ctx, id, updateUser)
+	ret := m.ctrl.Call(m, "Do", ctx, id, input)
 	ret0, _ := ret[0].(*app.UpdateUserByIdOutput)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // Do indicates an expected call of Do.
-func (mr *MockUpdateUserByIdInterfaceMockRecorder) Do(ctx, id, updateUser any) *gomock.Call {
+func (mr *MockUpdateUserByIdInterfaceMockRecorder) Do(ctx, id, input any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Do", reflect.TypeOf((*MockUpdateUserByIdInterface)(nil).Do), ctx, id, updateUser)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Do", reflect.TypeOf((*MockUpdateUserByIdInterface)(nil).Do), ctx, id, input)
 }