+++ CHANGELOG.md
@@ -10,4 +10,5 @@
 * BRAZ-9: Get user paginated Endpoint
 * BRAZ-34: Introduce logger with correlation Id
 * BRAZ-6: Update user Endpoint
+* BRAZ-35: Stop Bringing deleted rows
 * BRAZ-28: Introduce AWS and Terraform locally with Localstack